refactor(ERAS-152): start using the models created for the import
