+++ src/app/core/models/component.model.ts
@@ -1,7 +1,12 @@
 import { BaseModel } from './common/base.model';
-import { VariableModel } from './variable.model';
+import { Variable, VariableModel } from './variable.model';
 
 export interface ComponentModel extends BaseModel {
   name: string;
   variables: VariableModel[];
+}
+
+export interface Component extends BaseModel {
+  name: string;
+  variables: Variable[];
 }
