+++ src/app/core/models/variable.model.ts
@@ -1,3 +1,4 @@
+import { Answer } from './answer.model';
 import { BaseModel } from './common/base.model';
 
 export interface VariableModel extends BaseModel {
@@ -6,4 +7,11 @@ export interface VariableModel extends BaseModel {
   idComponent: number;
   pollVariableId: number;
   idPoll: number;
+}
+
+export interface Variable extends BaseModel {
+  name: string;
+  position: number;
+  type: string;
+  answer: Answer;
 }
